fixed the test cases
