--- frappe/core/page/data_import_tool/data_import_tool.json
@@ -1,19 +1,0 @@
-{
- "content": null, 
- "creation": "2012-06-14 15:07:25", 
- "docstatus": 0, 
- "doctype": "Page", 
- "icon": "fa fa-upload", 
- "idx": 1, 
- "modified": "2016-05-11 03:37:53.385693", 
- "modified_by": "Administrator", 
- "module": "Core", 
- "name": "data-import-tool", 
- "owner": "Administrator", 
- "page_name": "data-import-tool", 
- "roles": [], 
- "script": null, 
- "standard": "Yes", 
- "style": null, 
- "title": "Data Import Tool"
-}
